Handle when nodes are down
The following changes handle the scenario where the other nodes in
a HA setup has just gone. They might have OOM'd and now have a new
IP, or in the case of a cloud where they just kill the instance. This
will work out the ip of the unit that you're on and then attempt to
force that as the leader.

If for any reason the instances come back then you will cause a split
brain. It is recommended that all controller agents are stopped. You
can always bring back a node if it was flapping by removing the dqlite
folder in /var/lib/juju/dqlite and starting the controller agent again.
It will rejoin from where it left.